# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Surplus Food Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: RegisterRequestRole;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  registrationNumber?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
}