# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Surplus Food Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  organizationName?: string | null;
}